internal-fix(schemas): tighten ShardMetadata sample_rate type + AST-based leaf-import test
Addresses Copilot review on PR #976:

- ShardMetadata.sample_rate: float → int. The h5py audio attr is written
  from RenderConfig.sample_rate (int), so the wds sidecar mirrors the
  canonical type now rather than drifting at the format boundary. Test
  payloads updated to match.
- The leaf-module test now parses the module's AST and asserts no
  ImportFrom/Import nodes targeting src.* — replaces the substring grep,
  which would have false-failed on a docstring mentioning "from src." and
  missed alternative import phrasings.

Refs #975

+++ src/pipeline/schemas/shard_metadata.py
@@ -23,6 +23,6 @@ class ShardMetadata(BaseModel):
 
     velocity: int
     signal_duration_seconds: float
-    sample_rate: float
+    sample_rate: int
     channels: int
     min_loudness: float
